[#TBD] PG19: Fix Linux builds
Summary:
Odyssey's scram.c currently doesn't compile against PG19. Disable it for now.
With this change, pg19 branch compiles on Linux in both release and debug builds.

Test Plan: ./yb_build.sh release and ./yb_build.sh debug on alma8 devserver.

Reviewers: kramanathan

Differential Revision: https://phorge.dev.yugabyte.com/D54196

+++ src/postgres/src/backend/catalog/yb_catalog/yb_catalog_version.c
@@ -822,7 +822,8 @@ YbDeleteMasterDBInvalidationMessagesTableEntries(Oid db_oid)
 			elog(ERROR, "SPI_execute_plan failed for \"%s\"", query);
 		ereport((*YBCGetGFlags()->log_ysql_catalog_versions ? LOG : DEBUG1),
 				(errmsg("%s: deleted %llu invalidation messages for database %u",
-						__func__, SPI_processed, db_oid)));
+						__func__, (unsigned long long) SPI_processed,
+						db_oid)));
 	}
 	PG_CATCH();
 	{
